Fixed the problem with the FresnelLightingModel
Wasn't a problem with the FresnelLightingModel
Actually a problem with the Material for AIR -- which apparently had
been given a surface-transparency of 0.0 -- meaning that once a ray
refracted into AIR, that ray suddenly was given a color of NOTHING BUT
WHITE.

So everything was behaving precisely correctly. I was just giving it
garbage.