Builders can now be invoked automatically based on HasName values!
Given a tree of WorldFileObjectDefinitions, the BuilderInvoker will
execute appropriate Builder instances (and methods on those instances)
based on the names and values given in the ...ObjectDefinitions.

The NamedBuilderRegistrar leverages Spring to scan the classpath for
annotated Builder types and methods.

The next step would be to link WorldFileObjectDefinition generation to
.world-file scanning.