Patched hole which allowed unnoticed changes.
The emissive-shapes collection is calculated/derived from the general
Shapes list only once after that list has changed in any way. Hitherto,
that emissive-shapes collection was simply returned, which allowed dumb
mistakes to overwrite that list (which was then never re-calculated).

now, that emissive-shapes collection is deliberately converted to an
unmodifiable list prior to being returned.